advanced search for classes and links now work
Using the advanced search for classes and links should now work! users
and everything still do not work, that will require far more work.

Also the method of getting the class and link search to work is probably
horribly innificient, please let me know if it works and if there are
bugs.